Getting tests to run in Appveyor.
Updating init.ps1 to download, install, and configure an instance of Bitbucket Server so that it is ready to receive API requests. It generates an admin credential and exports it to .bbservercredential, in the root of the working directory. This file can be converted back to a PSCredential object using Import-CliXml.